backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "fastapi[standard]>=0.116.1",
    "langchain-core>=0.3.72",
    "langgraph>=0.6.3",
    "motor>=3.3.0",
    "openai>=1.98.0",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
]
